Fix for GetList
Currently the GetList() function for IssueLinkTypes in OnPremise and
Cloud fails to return the JSON document from the API. This is because it
expects a top level array of subdocuments and not a top level document
containing an array of subdocuments. I have corrected this behavior and
revised the tests to ensure the behavior is correct. I have also tested
this behavior against my own Jira OnPremise and Cloud instances.